Revert "fix: split use_veto_vote"
This reverts commit 197cd02483f1e79b5b8a94b94489162f16b04c92.
